Move over all the information from the .txt into a ods file as our artists requested. I tried to find proper fdo-names for the icons, but for some I don't even see in which category they would fit (the ones listed in the second part, already asked on kde-devel).
The text file should be kept for now as reference in case my comments in the
ods are unclear.

@KDevelop team: Please review and add as you see needed, also please think
about the needed sizes, for now I've added 16x16, 22x22 (those are the common
toolbar sizes and menus) as well as 32x32. I'm not sure about the latter for
all the things.

@Artists team: Let me know if there are any questions coming up or you need
anything else(IRC: apaku, email should be findable pretty easily). Note that
there's not yet any hurry, we're getting closer to a release but are still
looking at a couple of months of development at least. But we also didn't want
to wait until the last second. I'll see tomorrow wether the wiki page has
anything else thats not listed here. I'm subscribed to kde-artist now but won't
read the mails a lot, so better cc' me on any important things.

CCMAIL:kdevelop-devel@kdevelop.org
CCMAIL:kde-artists@kde.org


git-svn-id: svn://anonsvn.kde.org/home/kde/trunk/playground/artwork@979076 283d02a7-25f6-0310-bc7c-ecb5cbfe19da