Read emerald-style buttons images
This reads images from a single image file (one per button type),
containing six images for normal, hover, press states, and the
same for inactive windows. The inactive images are not yet used,
though, but you can prepare your themes now.

The files have to be renamed. For example:

    buttons.unsticky.png -> buttonsStickyDown.png

Every button type can have a different size, but it is required
that each state of the button has the same size, i.e. the image
width must be a multiple of 6.

Note that the old separate image files can still be provided for
compatibility with older deKorator versions, they are ignored
per button type when an emerald-style file is found.


git-svn-id: svn://anonsvn.kde.org/home/kde/trunk/playground/artwork@1108600 283d02a7-25f6-0310-bc7c-ecb5cbfe19da