Rename kplato -> plan + new app and mime icons
App and mime is in line 245 - 248
Generally I intended to use the same icons as for kplato as kplato is essentially dead.
Except they include the KOffice logo. Maybe just remove the logo?

Also icons for planwork is missing. This is a helper app for plan, essentially a todo list.
I was thinking of using the same icon as for plan with a "mark" like in the view-calendar-tasks icon.


git-svn-id: svn://anonsvn.kde.org/home/kde/trunk/playground/artwork@1271070 283d02a7-25f6-0310-bc7c-ecb5cbfe19da